requirements.txt
uuid
django
djangorestframework
coverage